feat(build) Update build script

+++ deploy.sh
@@ -2,12 +2,21 @@
 
 set -e
 
-echo '[LOG] Build application'
-yarn run build:optimized --base-href=/
-echo '[LOG] Clean previous packaged applications'
-rm -f devtools.zpush.io.tar.gz
-echo '[LOG] Package application'
-cd ./dist
-tar cvfz ../devtools.zpush.io.tar.gz ./**
-echo '[LOG] Deploy application on webserver'
-rsync -azv ./** ubuntu@devtools.zpush.io:/var/www/devtools.zpush.io/
+hash=$(git rev-parse --verify HEAD)
+artifact="devtools.zpush.io--$hash.tar.gz"
+file=./archives/$artifact
+
+if [ -f $file ]; then
+   echo "[LOG] Artifact already exists"
+else
+  echo "[LOG] Build application"
+  yarn run build:optimized --base-href=/
+  echo "[LOG] Clean previous packaged applications"
+  rm -f $file
+  echo "[LOG] Package application"
+  cd ./dist
+  echo "[LOG] Generate artifact $artifact"
+  tar cvfz .$file ./**
+  echo "[LOG] Deploy application on webserver"
+  rsync -azv ./** ubuntu@devtools.zpush.io:/var/www/devtools.zpush.io/
+fi
